Modularize folder breadcrumbs and flow table filters for improved reusability (#1238)
